This is the last bug, exceptions not raised if a directory is attempted to be deleted if not empty. Test added to verify that we don't delete said directories. Using dircache.listdir to check for empty.

+++ oggify/__init__.py
@@ -18,6 +18,7 @@
 # Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 
 import os, os.path, sys, re, tempfile, shutil, atexit
+from dircache import listdir
 from oggify import utils
 
 version = '2.0.0rc3'
@@ -134,7 +135,8 @@ class Oggify(object):
             print "Removing %s" % item
             if act:
                 if os.path.isdir(item):
-                    # TODO: Skip if directory not empty
+                    if len(listdir(item)) > 0:
+                        continue
                     os.removedirs(item)
                 else:
                     os.unlink(item)
